+++ src/components/reusable/dropdown/dropdown.ts
@@ -9,6 +9,7 @@ import {
 import { classMap } from 'lit/directives/class-map.js';
 import DropdownScss from './dropdown.scss';
 import './dropdownOption';
+import '../tag';
 import '@kyndryl-design-system/shidoka-foundation/components/icon';
 import downIcon from '@carbon/icons/es/chevron--down/24';
 import errorIcon from '@carbon/icons/es/warning--filled/24';
@@ -18,6 +19,7 @@ import clearIcon16 from '@carbon/icons/es/close/16';
 /**
  * Dropdown, single select.
  * @fires on-change - Captures the input event and emits the selected value and original event details.
+ * @fires on-search - Capture the search input event and emits the search text.
  * @slot unnamed - Slot for dropdown options.
  * @slot label - Slot for input label.
  */
@@ -331,7 +333,7 @@ export class Dropdown extends LitElement {
 
         ${this.multiple && !this.hideTags && this.value.length
           ? html`
-              <div class="tags">
+              <kyn-tag-group filter>
                 ${this.value.map((value: string) => {
                   const option = this.options.find(
                     (option) => option.value === value
@@ -351,17 +353,14 @@ export class Dropdown extends LitElement {
                   }
 
                   return html`
-                    <button
-                      class="tag"
-                      aria-label="Deselect ${text}"
-                      @click=${() => this.handleTagClear(option.value)}
-                    >
-                      ${text}
-                      <kd-icon .icon=${clearIcon16}></kd-icon>
-                    </button>
+                    <kyn-tag
+                      label=${text}
+                      @on-close=${(e: any) =>
+                        this.handleTagClear(e.detail.value)}
+                    ></kyn-tag>
                   `;
                 })}
-              </div>
+              </kyn-tag-group>
             `
           : null}
         ${this.caption !== ''
@@ -601,6 +600,8 @@ export class Dropdown extends LitElement {
     this.searchText = '';
     this.searchEl.value = '';
 
+    this._emitSearch();
+
     // clear selection for single select
     if (!this.multiple) {
       this.value = '';
@@ -660,6 +661,8 @@ export class Dropdown extends LitElement {
     const value = e.target.value;
     this.searchText = value;
     this.open = true;
+
+    this._emitSearch();
 
     // find matches
     const options = this.options.filter((option: any) => {
@@ -847,6 +850,15 @@ export class Dropdown extends LitElement {
     this.dispatchEvent(event);
   }
 
+  private _emitSearch() {
+    const event = new CustomEvent('on-search', {
+      detail: {
+        searchText: this.searchText,
+      },
+    });
+    this.dispatchEvent(event);
+  }
+
   override willUpdate(changedProps: any) {
     if (changedProps.has('open')) {
       if (this.open) {
@@ -940,8 +952,8 @@ export class Dropdown extends LitElement {
         }
 
         // set search input value
-        this.searchText = this.text === this.placeholder ? '' : this.text;
-        if (this.searchEl) {
+        if (this.searchable) {
+          this.searchText = this.text === this.placeholder ? '' : this.text;
           this.searchEl.value = this.searchText;
         }
       }